backend/requirements.txt
# Tile-Crawler Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# OpenAI Integration
openai==1.12.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Async Support
httpx==0.26.0
aiofiles==23.2.1

# CORS Support
starlette==0.35.1

# JSON Handling (included in stdlib, but explicit for type hints)
typing-extensions==4.9.0
